Update TF/SF source such that one can choose any combinations of the four edges.
Bug fix: the one corner radiating problem in TF/SF source.

Add waveguide modal TF/SF source.

Update docstring for ndc.

+++ tdyno/mnt/sig_spct.py
@@ -15,7 +15,7 @@ class SigSpct:
         ----------
         dt : float
             time step length
-        td : float
+        td : int
             time steps delay before starting to calculate the spectrum
         omin : float
             lower bound of omega
@@ -31,7 +31,7 @@ class SigSpct:
             number of signals in parallel
         """
         if dt is None:
-            dt = 1.
+            dt = 1
         self.dt = dt
         if td is None:
             td = 0.
